added remote server restart/re-connection
Every time the fragment gets destroyed the client now sends a 'restart'
signal to the ED_Tool-server to close the socket, open again and await
re-connection. the Client will then make sure the socket on its end is
closed. whenever that fragment is resumed or recreated the client-server
connection handshake is initialized so the re-connect is complete.
I have added this same process when opening the 'app settings', sockets
are closed, and reconnect takes place when that fragment is
reloaded/created